Split up texture module
