+++ demosys/opengl/texture/__init__.py
@@ -1,0 +1,3 @@
+from .texture import Texture2D
+from .array import TextureArray
+from .depth import DepthTexture
